refactor: move /server to /pkg/server
